requirements-ml.txt
# Parkinson's Disease Detection - ML Requirements
# =============================================
librosa>=0.9.0
scipy>=1.10.0

# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap>=0.41.0
matplotlib>=3.7.0
seaborn>=0.12.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
joblib>=1.3.0

# Data Processing
urllib3>=1.26.0

# For Model Server
Flask>=2.0.0
Flask-Cors>=3.0.0
fpdf2>=2.7.0

# Optional: For better performance
jupyter>=1.0.0
ipykernel>=6.25.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0